[#105] major refactoring; ;

+++ src/ml_service/feature_extraction/main.py
@@ -22,7 +22,7 @@ def cluster_means(arguments):
 def cluster_dbscan(arguments):
     data_tuple = get_precendece_model(arguments[0])
     start = time.time()
-    cluster_facts(data_tuple, int(arguments[1]), int(arguments[2]))
+    cluster_facts(data_tuple, int(arguments[1]), float(arguments[2]))
     done = time.time()
     Log.write('\nClustering time:')
     Log.write(done - start)
@@ -56,6 +56,7 @@ def parse_precedence(command):
         sys.exit(1)
     save_model.save(data)
 
+
 def process_command(command, arguments):
     if command == '--dbscan':
         cluster_dbscan(arguments)
@@ -69,9 +70,9 @@ def process_command(command, arguments):
         Log.write('Command not recognized:' + command)
         sys.exit(1)
 
+
 if __name__ == '__main__':
     if len(sys.argv) < 2:
         Log.write('Please provide more arguments')
         sys.exit(1)
-
     process_command(sys.argv[1], sys.argv[2:])
